requirements.txt
pygame
pillow
youtube-dl